feat(wallet): shared group switcher on desktop-remote + honest lock semantics
Desktop remote mode faked Core's timed-session lock model around the
BDK one-open-wallet reality. This unifies the switcher and makes every
lock surface tell the truth:

- new shared WalletGroupMenuComponent (extracted from the mobile shell):
  group rows with swipe rename/delete, create/restore entries, optional
  Manage Wallets top item - now ALSO the desktop toolbar's wallet menu
  in remote mode (Core keeps its load/unload menu). Closes the desktop
  rename/delete gap via the shared rows.
- locked btcx wallets are OPENABLE from desktop: select/open/load routes
  into the passphrase dialog (typed WALLET_LOCKED error from the
  manager) instead of a misleading "not loaded" toast or a silent
  console error; the padlock affordance shows for locked (unloaded)
  wallets on the wallet-select page and in the menus
- unlocking activates: on remote the unlocked wallet IS the open wallet,
  so all four unlock paths promote it to active (the toolbar no longer
  names the previous wallet)
- lock indicators: an OPEN group counts as unlocked - sibling
  compartments report seedLocked merely because their store isn't the
  open one (the "still locked after unlock" false positive, fixed in
  the summary mapping and the shared menu)
- Manage Wallets = VISIT (navigation-state flag passes noAuthGuard,
  wallet stays open and active - Core no longer loses its selection
  either); Logout on remote = close + lock (btcx_wallet_lock drops the
  runtime and the held passphrase)
- dev profile: scrypt/salsa20/sha2 at opt-level 3 - the at-rest KDF
  (N=2^15, by design expensive) made every unlock a multi-second stall
  in unoptimized dev builds
